+++ api/airport_data/__init__.py
@@ -1,0 +1,5 @@
+from flask import Flask, Blueprint
+
+airport_data = Blueprint("airport_data", __name__)
+
+import api.airport_data.routes
