+++ api/telex/__init__.py
@@ -1,0 +1,6 @@
+from flask import Flask, Blueprint
+
+telex = Blueprint("telex", __name__)
+
+import api.telex.models
+import api.telex.routes
