+++ run.py
@@ -3,4 +3,4 @@ from api import create_app
 app = create_app()
 
 if __name__ == "__main__":
-    app.run()
+    app.run(host='0.0.0.0')
